genai_solution/llm-cloudnative-demo/helm_chart/llm-deploy-chart/requirements.txt
interval==1.0.0
kubernetes==29.0.0
PyYAML==6.0.1
Requests==2.31.0
